+++ Source_code/Testing_pipeline/Load_model.py
@@ -4,7 +4,7 @@ import Config_Files.testing_pipeline_config as config
 
 import tensorflow as tf
 import os
-
+'''
 class LoadModel():
 
     def __init__(self, path):
@@ -23,4 +23,19 @@ class LoadModel():
             print("Entered Exception: Model could not be loaded.")
             print("Error is: {}".format(e))
 
+'''
 
+#global loaded_model
+
+def load_model():
+    #loaded_model= None
+    try:
+
+        print("Loading  the model")
+        trained_path = config.locations['trained_model_path']
+        load_model.loaded_model = tf.keras.models.load_model(trained_path)
+        print("Loaded the model")
+
+    except Exception as e:
+        print("Entered Exception: Model could not be loaded.")
+        print("Error is: {}".format(e))
